Finishing up functionality
